+++ plugins/blend-importer/src/index.ts
@@ -1,1 +1,2 @@
 export {BlendLoadPlugin} from './BlendLoadPlugin'
+export type {BlendFile, BlendLoadOptions} from './BlendLoadPlugin'